@@ -16,23 +16,16 @@
         }
     ],
     "require": {
-        "php": "^7.3",
+        "php": "^7.4|^8.0",
         "oauth2-framework/core": "^1.0",
         "sensio/framework-extra-bundle": "^5.0",
-        "symfony/config": "^4.3|^5.0",
-        "symfony/dependency-injection": "^4.3|^5.0",
-        "symfony/http-kernel": "^4.3|^5.0",
+        "symfony/config": "^5.0",
+        "symfony/dependency-injection": "^5.0",
+        "symfony/http-kernel": "^5.0",
         "symfony/psr-http-message-bridge": "^1.0|^2.0",
-        "symfony/routing": "^4.3|^5.0",
-        "symfony/security-bundle": "^4.3|^5.0",
+        "symfony/routing": "^5.0",
+        "symfony/security-bundle": "^5.0",
         "web-token/jwt-bundle": "^2.0"
-    },
-    "require-dev": {
-        "phpunit/phpunit": "^9.0",
-        "symfony/browser-kit": "^4.3|^5.0",
-        "symfony/framework-bundle": "^4.3|^5.0",
-        "symfony/phpunit-bridge": "^4.3|^5.0",
-        "symfony/yaml": "^4.3|^5.0"
     },
     "suggest": {
         "symfony/security-bundle": "Mandatory when resource owners can be end users."
@@ -40,11 +33,6 @@
     "autoload": {
         "psr-4": {
             "OAuth2Framework\\ServerBundle\\": ""
-        }
-    },
-    "extra": {
-        "branch-alias": {
-            "dev-master": "1.0.x-dev"
         }
     },
     "config": {
