@@ -51,7 +51,7 @@ class AuthorizationEndpointSource implements Component
     /**
      * @var Component[]
      */
-    private $subComponents = [];
+    private array $subComponents = [];
 
     public function __construct()
     {
