@@ -24,7 +24,7 @@ class RequestObjectSource implements Component
     /**
      * @var Component[]
      */
-    private $subComponents = [];
+    private array $subComponents = [];
 
     public function __construct()
     {
