@@ -27,7 +27,7 @@ class MetadataEndpointSource implements Component
     /**
      * @var Component[]
      */
-    private $subComponents = [];
+    private array $subComponents = [];
 
     public function __construct()
     {
