@@ -30,7 +30,7 @@ class OpenIdConnectSource implements Component
     /**
      * @var Component[]
      */
-    private $subComponents;
+    private array $subComponents;
 
     public function __construct()
     {
