@@ -26,7 +26,7 @@ class UserinfoEndpointSource implements Component
     /**
      * @var Component[]
      */
-    private $subComponents;
+    private array $subComponents;
 
     public function __construct()
     {
