@@ -21,10 +21,7 @@ use Symfony\Component\Routing\RouterInterface;
 
 final class ClientConfigurationRouteRule extends Base
 {
-    /**
-     * @var RouterInterface
-     */
-    private $router;
+    private RouterInterface $router;
 
     public function __construct(RouterInterface $router)
     {
