@@ -13,6 +13,8 @@ declare(strict_types=1);
 
 namespace OAuth2Framework\ServerBundle\Service;
 
+use function Safe\sprintf;
+use function Safe\parse_url;
 use Base64Url\Base64Url;
 use OAuth2Framework\Component\AuthorizationEndpoint\AuthorizationRequest\AuthorizationRequest;
 use OAuth2Framework\Component\OpenIdConnect\ConsentScreen\SessionStateParameterExtension as BaseSessionStateParameterExtension;
@@ -22,39 +24,15 @@ use Symfony\Component\HttpFoundation\Session\SessionInterface;
 
 class SessionStateParameterExtension extends BaseSessionStateParameterExtension
 {
-    /**
-     * @var string
-     */
-    private $storageName;
+    private string $storageName;
 
-    /**
-     * @var SessionInterface
-     */
-    private $session;
-    /**
-     * @var null|string
-     */
-    private $path;
-    /**
-     * @var null|string
-     */
-    private $domain;
-    /**
-     * @var bool
-     */
-    private $secure;
-    /**
-     * @var bool
-     */
-    private $httpOnly;
-    /**
-     * @var bool
-     */
-    private $raw;
-    /**
-     * @var null|string
-     */
-    private $sameSite;
+    private SessionInterface $session;
+    private ?string $path;
+    private ?string $domain;
+    private bool $secure;
+    private bool $httpOnly;
+    private bool $raw;
+    private ?string $sameSite;
 
     public function __construct(SessionInterface $session, string $storageName, ?string $path = '/', ?string $domain = null, bool $secure = false, bool $httpOnly = true, bool $raw = false, ?string $sameSite = null)
     {
@@ -87,15 +65,15 @@ class SessionStateParameterExtension extends BaseSessionStateParameterExtension
         $redirectUri = $authorization->getRedirectUri();
         $origin = $this->getOriginUri($redirectUri);
         $salt = Base64Url::encode(random_bytes(16));
-        $hash = hash('sha256', \Safe\sprintf('%s%s%s%s', $authorization->getClient()->getPublicId(), $origin, $browserState, $salt));
+        $hash = hash('sha256', sprintf('%s%s%s%s', $authorization->getClient()->getPublicId(), $origin, $browserState, $salt));
 
-        return \Safe\sprintf('%s.%s', $hash, $salt);
+        return sprintf('%s.%s', $hash, $salt);
     }
 
     private function getOriginUri(string $redirectUri): string
     {
-        $url_parts = \Safe\parse_url($redirectUri);
+        $url_parts = parse_url($redirectUri);
 
-        return \Safe\sprintf('%s://%s%s', $url_parts['scheme'], $url_parts['host'], isset($url_parts['port']) ? ':'.$url_parts['port'] : '');
+        return sprintf('%s://%s%s', $url_parts['scheme'], $url_parts['host'], isset($url_parts['port']) ? ':'.$url_parts['port'] : '');
     }
 }
