@@ -18,15 +18,9 @@ use OAuth2Framework\Component\MacTokenType\MacToken as Base;
 
 final class MacToken extends Base
 {
-    /**
-     * @var int
-     */
-    private $minLength;
+    private int $minLength;
 
-    /**
-     * @var int
-     */
-    private $maxLength;
+    private int $maxLength;
 
     public function __construct(string $macAlgorithm, int $timestampLifetime, int $minLength, int $maxLength)
     {
