@@ -13,6 +13,7 @@ declare(strict_types=1);
 
 namespace OAuth2Framework\ServerBundle\Component\TokenType;
 
+use function Safe\sprintf;
 use OAuth2Framework\Component\Core\TokenType\TokenTypeManager;
 use Symfony\Component\DependencyInjection\Compiler\CompilerPassInterface;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
@@ -34,7 +35,7 @@ class TokenTypeCompilerPass implements CompilerPassInterface
         foreach ($taggedServices as $id => $tags) {
             foreach ($tags as $attributes) {
                 if (!\array_key_exists('scheme', $attributes)) {
-                    throw new \InvalidArgumentException(\Safe\sprintf('The token type "%s" does not have any "scheme" attribute.', $id));
+                    throw new \InvalidArgumentException(sprintf('The token type "%s" does not have any "scheme" attribute.', $id));
                 }
                 $is_default = $default === $attributes['scheme'];
                 $token_type_names[] = $attributes['scheme'];
@@ -46,7 +47,7 @@ class TokenTypeCompilerPass implements CompilerPassInterface
         }
 
         if (!$default_found) {
-            throw new \InvalidArgumentException(\Safe\sprintf('Unable to find the token type "%s". Available token types are: %s.', $default, implode(', ', $token_type_names)));
+            throw new \InvalidArgumentException(sprintf('Unable to find the token type "%s". Available token types are: %s.', $default, implode(', ', $token_type_names)));
         }
     }
 }
