@@ -1,6 +1,16 @@
 <?php
 
 declare(strict_types=1);
+use OAuth2Framework\Component\Core\Middleware\OAuth2MessageMiddleware;
+use OAuth2Framework\Component\Core\Message\OAuth2MessageFactoryManager;
+use OAuth2Framework\Component\Core\Message\Factory\AccessDeniedResponseFactory;
+use OAuth2Framework\Component\Core\TokenType\TokenTypeManager;
+use OAuth2Framework\Component\Core\Message\Factory\AuthenticateResponseForTokenFactory;
+use OAuth2Framework\Component\Core\Message\Factory\AuthenticateResponseForClientFactory;
+use OAuth2Framework\Component\Core\Message\Factory\BadRequestResponseFactory;
+use OAuth2Framework\Component\Core\Message\Factory\MethodNotAllowedResponseFactory;
+use OAuth2Framework\Component\Core\Message\Factory\NotImplementedResponseFactory;
+use OAuth2Framework\Component\Core\Message\Factory\RedirectResponseFactory;
 
 /*
  * The MIT License (MIT)
@@ -26,26 +36,26 @@ return function (ContainerConfigurator $container) {
     ;
 
     $container->set('oauth2_server.message_middleware.for_client_authentication')
-        ->class(Middleware\OAuth2MessageMiddleware::class)
+        ->class(OAuth2MessageMiddleware::class)
         ->args([
             ref('oauth2_server.message_factory_manager.for_client_authentication'),
         ])
     ;
     $container->set('oauth2_server.message_factory_manager.for_client_authentication')
-        ->class(Message\OAuth2MessageFactoryManager::class)
+        ->class(OAuth2MessageFactoryManager::class)
         ->args([
             ref(ResponseFactoryInterface::class),
         ])
     ;
 
     $container->set('oauth2_server.message_middleware.for_token_authentication')
-        ->class(Middleware\OAuth2MessageMiddleware::class)
+        ->class(OAuth2MessageMiddleware::class)
         ->args([
             ref('oauth2_server.message_factory_manager.for_token_authentication'),
         ])
     ;
     $container->set('oauth2_server.message_factory_manager.for_token_authentication')
-        ->class(Message\OAuth2MessageFactoryManager::class)
+        ->class(OAuth2MessageFactoryManager::class)
         ->args([
             ref(ResponseFactoryInterface::class),
         ])
@@ -53,16 +63,16 @@ return function (ContainerConfigurator $container) {
 
     //Factories
     $container->set('oauth2_server.message_factory.403')
-        ->class(Message\Factory\AccessDeniedResponseFactory::class)
+        ->class(AccessDeniedResponseFactory::class)
         ->tag('oauth2_server_message_factory_for_token_authentication')
         ->tag('oauth2_server_message_factory_for_client_authentication')
     ;
 
     $container->set('oauth2_server.message_factory.401_for_token')
         ->args([
-            ref(\OAuth2Framework\Component\Core\TokenType\TokenTypeManager::class),
+            ref(TokenTypeManager::class),
         ])
-        ->class(Message\Factory\AuthenticateResponseForTokenFactory::class)
+        ->class(AuthenticateResponseForTokenFactory::class)
         ->tag('oauth2_server_message_factory_for_token_authentication')
     ;
 
@@ -70,30 +80,30 @@ return function (ContainerConfigurator $container) {
         ->args([
             ref(AuthenticationMethodManager::class),
         ])
-        ->class(Message\Factory\AuthenticateResponseForClientFactory::class)
+        ->class(AuthenticateResponseForClientFactory::class)
         ->tag('oauth2_server_message_factory_for_client_authentication')
     ;
 
     $container->set('oauth2_server.message_factory.400')
-        ->class(Message\Factory\BadRequestResponseFactory::class)
+        ->class(BadRequestResponseFactory::class)
         ->tag('oauth2_server_message_factory_for_token_authentication')
         ->tag('oauth2_server_message_factory_for_client_authentication')
     ;
 
     $container->set('oauth2_server.message_factory.405')
-        ->class(Message\Factory\MethodNotAllowedResponseFactory::class)
+        ->class(MethodNotAllowedResponseFactory::class)
         ->tag('oauth2_server_message_factory_for_token_authentication')
         ->tag('oauth2_server_message_factory_for_client_authentication')
     ;
 
     $container->set('oauth2_server.message_factory.501')
-        ->class(Message\Factory\NotImplementedResponseFactory::class)
+        ->class(NotImplementedResponseFactory::class)
         ->tag('oauth2_server_message_factory_for_token_authentication')
         ->tag('oauth2_server_message_factory_for_client_authentication')
     ;
 
     $container->set('oauth2_server.message_factory.303')
-        ->class(Message\Factory\RedirectResponseFactory::class)
+        ->class(RedirectResponseFactory::class)
         ->tag('oauth2_server_message_factory_for_token_authentication')
         ->tag('oauth2_server_message_factory_for_client_authentication')
     ;
